Addresses #152
For some reason, although the atom index orders are maintained, the bond index orders are not always maintained in RDKit object creation, even if bonds have been specified in the SDF file. This problem has been addressed by comparing the atom indices shown within bonds of RDKit and the atoms in the original molli object.