`jobmap` update
The goal of this commit is to wrap the full functionality of the paralellization into a unified class now referred to as the `Jobmap` class.  This not only eliminates a lot of repetitve code between `jobmap` and `jobmap_sge`, but it isolates the intermediate step of job submissions locally or to HPCs as a separate function.  This should enable the addition of different cluster configurations to the class in the future.

Some notable aspects of this new class  include:

- Instantiation results in creation of cache directory and input files
- Functionality has been adapted for both local and sge-configured HPCs.
- Output collection is separated from the individual submission and monitoring of jobs.
- A majority of the code run through `jobmap` or `jobmap_sge` have not been altered/modified, rather just wrapped to reduce reptition.